+++ maui5eClient/Extensions/RegisterFontsExtension.cs
@@ -1,4 +1,4 @@
-namespace maui5e.Extensions;
+namespace maui5eClient.Extensions;
 
 public static class RegisterFontsExtension
 {
