+++ maui5eClient/Extensions/RegisterHandlersExtension.cs
@@ -1,4 +1,4 @@
-namespace maui5e.Extensions;
+namespace maui5eClient.Extensions;
 
 public static class RegisterHandlersExtension
 {
