+++ maui5eClient/Platforms/MacCatalyst/Program.cs
@@ -1,7 +1,7 @@
 ﻿using ObjCRuntime;
 using UIKit;
 
-namespace maui5e;
+namespace maui5eClient;
 
 public class Program
 {
