+++ maui5eClient/maui5eClient.App.sln
