+++ maui5eClient/maui5eClient.csproj
@@ -6,17 +6,19 @@
         <!-- Uncomment to also build the tizen app. You will need to install tizen by following this: https://github.com/Samsung/Tizen.NET -->
         <!-- <TargetFrameworks>$(TargetFrameworks);net6.0-tizen</TargetFrameworks> -->
         <OutputType>Exe</OutputType>
-        <RootNamespace>maui5e</RootNamespace>
+        <RootNamespace>maui5eClient</RootNamespace>
         <UseMaui>true</UseMaui>
         <SingleProject>true</SingleProject>
         <ImplicitUsings>enable</ImplicitUsings>
+        <SkipValidateMauiImplicitPackageReferences>true</SkipValidateMauiImplicitPackageReferences>
+
 
         <AcceptAndroidSDKLicenses>true</AcceptAndroidSDKLicenses>
         <!-- Display name -->
-        <ApplicationTitle>maui5e</ApplicationTitle>
+        <ApplicationTitle>maui5eClient</ApplicationTitle>
 
         <!-- App Identifier -->
-        <ApplicationId>com.companyname.maui5e</ApplicationId>
+        <ApplicationId>com.companyname.maui5eClient</ApplicationId>
         <ApplicationIdGuid>A837A198-63DB-4B57-9AC3-66BD9939DFC8</ApplicationIdGuid>
 
         <!-- Versions -->
@@ -53,10 +55,6 @@
       <PackageReference Include="CommunityToolkit.Maui.Markup" Version="4.1.0" />
       <PackageReference Include="GraphQL.Client" Version="6.1.0" />
       <PackageReference Include="GraphQL.Client.Serializer.SystemTextJson" Version="6.1.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Folder Include="Data\" />
     </ItemGroup>
 
 </Project>
